✨ Feature: 필터링 유지 기능 추가 및 리팩토링
